requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.1
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
itsdangerous==2.2.0
python-dateutil==2.9.0.post0
alembic==1.13.2
pytest==8.3.2
